requirements.txt
APScheduler==3.11.0
fastapi==0.115.12
gpiozero==2.0.1
isbnlib==3.10.14
jose==1.0.0
motor==3.7.0
ndeflib==0.3.3
nfcpy==1.0.4
numpy==1.24.2
opencv_python==4.11.0.86
picamera2==0.3.25
Pillow==11.2.1
pycountry==24.6.1
pydantic==2.11.3
PyJWT==2.10.1
pymongo==4.12.0
pynfc==0.1.1
python_bcrypt==0.3.2
python_jose==3.4.0
pyzbar==0.1.9
rapidfuzz==3.13.0
Requests==2.32.3
uvicorn==0.34.2